% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_doc.R
\docType{data}
\name{agents}
\alias{agents}
\title{ICEWS CAMEO actor codes}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 614 rows and 3 columns.}
\usage{
agents
}
\description{
Merge table to convert entries to CAMEO format using conversion tables created
   by Phil Schrodt (https://github.com/philip-schrodt/text_to_CAMEO)
}
\details{
@docType data

@usage data(agents)

@keywords datasets
}
\keyword{datasets}

