% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_icews.R
\name{get_icewslinks}
\alias{get_icewslinks}
\title{Download the ICEWS Dataset}
\usage{
get_icewslinks()
}
\arguments{
\item{destpath}{The path to the directory where ICEWS should go.}
}
\description{
Download and unzip all of the data files for the ICEWS dataset from the
Harvard Dataverse into a given directory.
}
\note{
This function is still in development and may contain errors and change quickly.
}
\examples{

download_icews("~/ICEWS/")

}
\author{
Original code and concept: Tony Boyles
}

