% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_dyadstats.R
\name{extract_dyadstats}
\alias{extract_dyadstats}
\title{Extract dyad-level statistics from a given event-network.}
\usage{
extract_dyadstats(input_date = this_date, input_net = tsna_obj)
}
\arguments{
\item{net_obj}{network object object containing a set of interactions.}
}
\value{
net_stats Table of dyad-level statistics.
}
\description{
INTERNAL FUNCTION: Intakes a given network object and returns a set
 of dyad-level statistics for output.
}
\keyword{data}
\keyword{event}
\keyword{phoenix,}

