% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nodestats.R
\name{extract_nodestats}
\alias{extract_nodestats}
\title{Extract node-level statistics from a given event-network.}
\usage{
extract_nodestats(input_date = this_date, event_dnet = tsna_obj)
}
\description{
INTERNAL FUNCTION: Intakes a given network object and returns a set
 of node-level statistics for output.
}
\details{
@param input_date A date in integer %Y%m%d format.
 @param event_dnet network object object containing a set of interactions.

@return net_stats Table of node-level statistics.

@keywords phoenix, event data

@import data.table
 @import countrycode
 @import reshape2
 @import statnet
 @import tsna
 @import plyr
 @import lubridate
 @import igraph
 @import intergraph

@export
}

