% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/icews_cameo.R
\name{icews_cameo}
\alias{icews_cameo}
\title{Convert ICEWS state/actor codes into CAMEO format,
 and extract root codes from specific CAMEO event codes.}
\usage{
icews_cameo(icews)
}
\arguments{
\item{icews}{ICEWS data as one large data.table}
}
\value{
icews ICEWS data with several new CAMEO code columns.
}
\description{
Intake a set of ICEWS data (read in after some pre-processing)
   and convert entries to CAMEO format using conversion tables created
   by Phil Schrodt (https://github.com/philip-schrodt/text_to_CAMEO)
}
\keyword{data}
\keyword{event}
\keyword{phoenix,}

