% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_histphoenix.R
\name{ingest_histphoenix}
\alias{ingest_histphoenix}
\title{Ingest the historic Phoenix Dataset}
\usage{
ingest_histphoenix(histphoenix_loc, start_date = start_date,
  end_date = end_date, statelist = statelist)
}
\arguments{
\item{histphoenix_loc}{The path to the Phoenix folder.}

\item{end_date}{}
}
\value{
A single dataframe with all the historic Phoenix events in the folder.
}
\description{
Given a directory with the historic Phoenix dataset files, quickly read
them all in, name them correctly, and combine them into one dataframe.
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{

events <- ingest_histphoenix("~/histphoenix")

}
