% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_icews.R
\name{ingest_icews}
\alias{ingest_icews}
\title{Ingest the ICEWS Event Dataset}
\usage{
ingest_icews(dir, start_date, end_date)
}
\arguments{
\item{dir}{The path to the ICEWS folder.}

\item{start_date}{Start of date range as YYYYMMDD integer format.}

\item{end_date}{End of date range as YYYYMMDD integer format.}
}
\value{
A single dataframe with all the ICEWS events in the folder.
}
\description{
Given a directory with individual ICEWS dataset files, quickly read
them all in, name them correctly, and combine them into one dataframe.
}
\note{
This function is still in development and may contain errors and change quickly.
}
\examples{

events <- ingest_icews("~/ICEWS/study_28075/Data/", 20101201, 20140101)

}
\author{
Andy Halterman, forked by Jesse Hammond
}

