% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_phoenix.R
\name{ingest_phoenix}
\alias{ingest_phoenix}
\title{Ingest the Phoenix Dataset}
\usage{
ingest_phoenix(phoenix_loc, start_date, end_date)
}
\arguments{
\item{phoenix_loc}{The path to the Phoenix folder.}

\item{start_date}{Start of date range as YYYYMMDD integer format.}

\item{end_date}{End of date range as YYYYMMDD integer format.}
}
\value{
A single dataframe with all the Phoenix events in the folder.
}
\description{
Given a directory with individual Phoenix dataset files, quickly read
them all in, name them correctly, and combine them into one dataframe.
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{

events <- ingest_phoenix("~/OEDA/phoxy_test/", 20140620, 20150101)

}
\author{
Andy Halterman, forked by Jesse Hammond
}

