% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phoenix_stats.R
\name{phoenix_stats}
\alias{phoenix_stats}
\title{Extract statistics from daily Phoenix event-networks.}
\usage{
phoenix_stats(dailynets)
}
\arguments{
\item{dailynets}{networkDynamic object containing daily event-nets
         produced via phoenix_net function.}
}
\value{
phoenix_out a LIST object of tables containing descriptive
         statistics for daily event-networks.
}
\description{
Take a set of daily networks generated by the phoenix_net function,
 and extract a variety of daily statistics at the network and nodal
 levels.
}
\keyword{data}
\keyword{event}
\keyword{phoenix,}

