% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoenix_tables.R
\name{phoenix_tables}
\alias{phoenix_tables}
\title{Scrape, merge, and process Phoenix and ICEWS data into
a large data table for aggregation and subsetting.}
\usage{
phoenix_tables(phoenix_loc, icews_loc, update = T)
}
\arguments{
\item{phoenix_loc}{folder containing Phoenix data sets as daily .csv
data tables. Automatically checks for new data sets each time
the function is run, and downloads new daily data as it becomes
available. Currently in 'one-and'done' format
where it downloads the first time, and checks thereafter.}

\item{icews_loc}{folder containing ICEWS data sets as daily .tab data
tables. Because I don't know how to work a SWORD API, these will
need to be manually downloaded and updated.}
}
\value{
master_table a data.table object containing ALL merged/processed
         Phoenix and ICEWS data. One row per event-dyad-day.
}
\description{
Scrape, merge, and process Phoenix and ICEWS data into
a large data table for aggregation and subsetting.
}
\note{
This function is still in early development and may contain significant errors.
       Don't trust it.
}
\author{
Jesse Hammond
}

