% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_doc.R
\docType{data}
\name{states}
\alias{states}
\title{ICEWS CAMEO actor codes}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 260 rows and 3 columns.}
\usage{
data(states)
}
\description{
Merge table to convert states to CAMEO format using conversion tables created
   by Phil Schrodt (https://github.com/philip-schrodt/text_to_CAMEO)
}
\keyword{datasets}

