% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_icews.R
\name{update_icews}
\alias{update_icews}
\title{Update a local directory of ICEWS dataset files with new files from the server}
\usage{
update_icews(destpath)
}
\arguments{
\item{destpath}{The path to download ICEWS into.}
}
\description{
Checks the contents of a directory containing ICEWS event data files, checks whether the
server has new events, and downloads them to that directory. (It'll have some version handling ability,
too, either from the file names or by reading in the events.)
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{

}
\author{
Original concept and code for Phoenix: Andy Halterman
}

