% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_phoenix.R
\name{get_phoenixlinks}
\alias{get_phoenixlinks}
\title{Download the Phoenix Dataset}
\usage{
get_phoenixlinks(start_date = as.Date("2014-06-20"), end_date = Sys.Date())
}
\arguments{
\item{destpath}{The path to the directory where Phoenix should go.}

\item{phoenix_version.}{Download a specific version of Phoenix ("v0.1.0" or the current version by default).}
}
\description{
Download and unzip all of the data files for the Phoenix dataset from the
Phoenix data website into a given directory.
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{

download_phoenix("~/OEDA/phoxy_test/", phoenix_version = "current")

}
\author{
Original code credit: Andy Halterman
}

