% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_netstats.R
\name{extract_netstats}
\alias{extract_netstats}
\title{Extract network-level statistics from a given event-network.}
\usage{
extract_netstats(input_date = this_date, event_dnet = event_dnet,
  datelist = dates)
}
\arguments{
\item{input_date}{A date in integer %Y%m%d format.}

\item{event_dnet}{network object object containing a set of interactions.}
}
\value{
net_stats Table of network-level statistics.
}
\description{
INTERNAL FUNCTION: Intakes a given network object and returns a set
 of network-level statistics for output.
}
\keyword{data}
\keyword{event}
\keyword{phoenix,}

