% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icews_cameo.R
\name{icews_cameo}
\alias{icews_cameo}
\title{Convert ICEWS state/actor codes into CAMEO format,
 and extract root codes from specific CAMEO event codes.}
\usage{
icews_cameo(icews)
}
\description{
Intake a set of ICEWS data (read in after some pre-processing)
   and convert entries to CAMEO format using conversion tables created
   by Phil Schrodt (https://github.com/philip-schrodt/text_to_CAMEO)
}
\details{
@param icews ICEWS data as one large data.table

@return icews ICEWS data with several new CAMEO code columns.

@keywords phoenix, event data

@import data.table
 @import plyr

@export
}

