% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phoenixNet.R
\name{phoenix_net}
\alias{phoenix_net}
\title{Convert Phoenix event data to daily event-networks.}
\usage{
phoenix_net(start_date, end_date, level)
}
\arguments{
\item{start_date}{start date of time period as Ymd-format integer (ex: June 1, 2014 as 20140601)}

\item{end_date}{end date of time period as Ymd-format integer (ex: June 1, 2014 as 20140601)}

\item{level}{level of event granularity ('eventcode' or 'rootcode'). 'Eventcode' creates
         a network for each of the 226 sub-codes in CAMEO. 'Rootcode' creates a network
         for each of the 20 event root codes in CAMEO.}
}
\value{
master_networks a LIST object containing daily event-networks.
}
\description{
Take event-level data and convert it into
 networks of interaction by time period. Output is in
 the form of a nested list object where each element is
 an R network object. These networks can then be processed
 and analyzed.
}
\keyword{data}
\keyword{event}
\keyword{phoenix,}

