% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoenix_net.R
\name{phoenix_net}
\alias{phoenix_net}
\title{Convert Phoenix event data to daily event-networks.}
\usage{
phoenix_net(start_date, end_date, level, phoenix_loc, icews_loc,
  update = TRUE, actorset = "states", codeset = "all",
  time_window = "day", code_subset = "all", tie_type = "binary",
  sources = "both")
}
\arguments{
\item{start_date}{start date of time period as Ymd-format integer (ex:
June 1, 2014 as 20140601).}

\item{end_date}{end date of time period as Ymd-format integer (ex:
June 1, 2014 as 20140601).}

\item{level}{level of event granularity ('eventcode', 'rootcode', or
'pentaclass). 'Eventcode' creates a network for each of the
226 sub-codes in CAMEO. 'Rootcode' creates a network for each
of the 20 event root codes in CAMEO. 'Pentaclass' creates a
network for each of the 0-4 pentaclass codes in CAMEO.}

\item{phoenix_loc}{folder containing Phoenix data sets as daily .csv
data tables. Automatically checks for new data sets each time
the function is run, and downloads new daily data as it becomes
available. Currently in 'one-and'done' format
where it downloads the first time, and checks thereafter.}

\item{icews_loc}{folder containing ICEWS data sets as daily .tab data
tables. Because I don't know how to work a SWORD API, these will
need to be manually downloaded and updated.}

\item{update}{should phoenixNet attempt to download new data? This will attempt
to download any Phoenix data files that 'should' be present in the
Phoenix data directory (one data file per day, from 2014-06-20 through
the present day) and denote whether or not any of these files
come up missing in the process.}

\item{actorset}{set of actors for which to create event-networks. Defaults
to the 255 ISO-coded states in the international system. Specifying
a specific state or set of states (as 3-character ISO codes) will
extract all the 'major' actors withint that state/states.}

\item{codeset}{subset of event codes as specified by 'level'. This is useful
if you desire to extract only a portion of interactions recorded
by CAMEO, but has to align with the code aggregation specified
in the 'level' argument. For example, if you specify 'rootcode',
the 'codeset' you specify has to be one or more root codes between
1 and 20. Defaults to 'all'.}

\item{time_window}{temporal window to build event-networks. Valid
entries are 'day', 'week', 'month', or 'year'.}

\item{code_subset}{subset of EVENTCODES that can be aggregated up to higher
order interactions. For example, you might want to only look at
event codes below 100, but then aggregate those event codes to
rootcode or pentaclass.}

\item{tie_type}{type of ties to return. Default is binarized ties where
a tie represents the presence of one OR MORE interactions in the
time period specified. Valid entries are 'binary', 'count'
(count of events), or 'pct' (percentage of total events between
the two actors). Currently only 'binary' and 'count' implemented.}

\item{sources}{use only Phoenix or ICEWS data in creating event networks.
Valid entries are 'phoenix', 'icews', or 'both' (default).}
}
\value{
master_networks a LIST object containing temporally referenced event-networks.
}
\description{
Take event-level data and convert it into
 networks of interaction by time period. Output is in
 the form of a nested list object where each element is
 an R network object. These networks can then be processed
 and analyzed.
}
\note{
This function is still in early development and may contain significant errors.
       Don't trust it.
}
\author{
Jesse Hammond
}

