% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoenix_stats.R
\name{phoenix_stats}
\alias{phoenix_stats}
\title{Extract statistics from daily Phoenix event-networks.}
\usage{
phoenix_stats(dailynets)
}
\description{
Take a set of daily networks generated by the phoenix_net function,
 and extract a variety of daily statistics at the network and nodal
 levels.
}
\details{
@param dailynets networkDynamic object containing daily event-nets
         produced via phoenix_net function.

@return phoenix_out a LIST object of tables containing descriptive
         statistics for daily event-networks.

@keywords phoenix, event data

@import data.table
 @import countrycode
 @import reshape2
 @import statnet
 @import tsna
 @import plyr
 @import lubridate
 @import igraph
 @import intergraph

@export
}

