% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoenix_stats.R
\name{phoenix_stats}
\alias{phoenix_stats}
\title{Extract statistics from daily Phoenix event-networks.}
\usage{
phoenix_stats(dailynets, time_window = "day", codes = "all",
  do_parallel = F, n_cores = 4)
}
\arguments{
\item{dailynets}{networkDynamic object containing daily event-nets
produced via phoenix_net function.}

\item{time_window}{time interval of aggregate event-network objects. Valid
entries are 'day', 'week', 'month', 'year'.}

\item{codes}{string of event codes, root codes, or pentaclass codes.
Note: these codes have to be in the same format as the original
network layers created via 'phoenix_net'. If you specify rootcodes
in the creation step, entering pentaclass codes in this step
will produce an error.}

\item{do_parallel}{Logical TRUE-FALSE. Whether to use parallel backend
'doMC' when extracting network statistics by code. Considerably
faster than single-core, but less reliable.}
}
\value{
phoenix_out a LIST object of tables containing descriptive
         statistics for daily event-networks.
}
\description{
Take a set of daily networks generated by the phoenix_net function,
 and extract a variety of daily statistics at the network and nodal
 levels.
}
\keyword{data}
\keyword{event}
\keyword{phoenix,}

